+++ src/widgets/starRating/types/types.ts
@@ -1,0 +1,6 @@
+export interface StarIconProps {
+  filled: boolean;
+  onClick: () => void;
+  onMouseEnter: () => void;
+  onMouseLeave: () => void;
+}
