+++ src/widgets/starRating/ui/StarRating.tsx
@@ -1,45 +1,7 @@
 import { useState } from 'react';
-import StarEmpty from '@shared/assets/images/star-rating/star.svg';
-import StarFilled from '@shared/assets/images/star-rating/star-filled.svg';
-import { create } from 'zustand';
-import { persist, createJSONStorage } from 'zustand/middleware';
+import { useRatingStore } from '../model/store';
+import { StarIcon } from './StarIcon';
 import styles from './StarRating.module.scss';
-
-interface RatingStore {
-  rating: number;
-  setRating: (rating: number) => void;
-}
-
-const useRatingStore = create<RatingStore>()(
-  persist(
-    (set) => ({
-      rating: 0,
-      setRating: (rating) => set({ rating }),
-    }),
-    {
-      name: 'rating-storage',
-      storage: createJSONStorage(() => localStorage),
-    }
-  )
-);
-
-interface StarIconProps {
-  filled: boolean;
-  onClick: () => void;
-  onMouseEnter: () => void;
-  onMouseLeave: () => void;
-}
-
-const StarIcon = ({ filled, onClick, onMouseEnter, onMouseLeave }: StarIconProps) => (
-  <img 
-    src={filled ? StarFilled : StarEmpty}
-    alt="star"
-    className={styles.star}
-    onClick={onClick}
-    onMouseEnter={onMouseEnter}
-    onMouseLeave={onMouseLeave}
-  />
-);
 
 const StarRating = () => {
   const { rating, setRating } = useRatingStore();
